requirements.txt
python-telegram-bot==13.15
requests==2.28.1
six==1.16.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
